+++ app/model/filters/filter_nagios_expr.py
@@ -17,24 +17,21 @@
 #   along with Sageo.  If not, see <http://www.gnu.org/licenses/>
 
 
-from .filter import Filter
+from .filter_tristate import FilterTristate
 from sqlalchemy import *
 from sqlalchemy.orm import *
 
-class FilterNagiosExpr(Filter):
-    def __init__(self, name, title, descr, pos_filter, neg_filter, default, column_names):
-        Filter.__init__(self, name, title, descr)
+class FilterNagiosExpr(FilterTristate):
+    def __init__(self, name, title, descr, default, column, pos_filter, neg_filter):
+        FilterTristate.__init__(self, name, title, descr, default, column)
         self.pos_filter = pos_filter
         self.neg_filter = neg_filter
-        self.default = default
-        self.column_names = column_names
 
     def filter(self, value):
-        if value == 'yes':
+        if value == '1':
             return self.pos_filter 
-        else:
+        elif value == '0': 
             return self.neg_filter 
-
-    def get_col_def(self):
-        return [Column(self.name, Enum('yes', 'no', 'ignore'), default=self.default)]
+        else:
+            return ""